tools/testing/connection/python-redis/redis-om/requirements.txt
redis-om>=0.2.0
redis>=4.0.0
python-dotenv
